% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ang_sep.R
\name{ang_sep}
\alias{ang_sep}
\title{Angular separation on the sphere}
\usage{
ang_sep(lon1, lat1, lon2, lat2, unit = c("rad", "deg"))
}
\arguments{
\item{lon1, lat1}{First point.}

\item{lon2, lat2}{Second point.}

\item{unit}{"rad" or "deg".}
}
\value{
Numeric vector.
}
\description{
Angular separation on the sphere
}
