% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astrocoords-package.R
\docType{package}
\name{astrocoords-package}
\alias{astrocoords}
\alias{astrocoords-package}
\title{astrocoords: Astronomical Coordinates for R}
\description{
Vectorized astronomical coordinates for R.
}
\author{
\strong{Maintainer}: Your Name \email{your@email.com}

}
\keyword{internal}
