% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{b}
\alias{b}
\title{Get Galactic latitude (b)}
\usage{
b(x)
}
\arguments{
\item{x}{A <sky_coord> vector in Galactic frame.}
}
\value{
Numeric vector with Galactic latitude values in degrees.
}
\description{
Extract Galactic latitude values (in degrees) from Galactic coordinates.
}
\examples{
x <- gal_coord(c(120, 121), c(-30, -31))
b(x)
}
