% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexagesimal.R
\name{dms_to_deg}
\alias{dms_to_deg}
\title{Convert DMS components to degrees}
\usage{
dms_to_deg(d, m = 0, s = 0)
}
\arguments{
\item{d}{Degrees (signed).}

\item{m}{Arcminutes.}

\item{s}{Arcseconds.}
}
\value{
Numeric vector in degrees.
}
\description{
Convert degree-minute-second components to decimal degrees.
}
\examples{
dms_to_deg(-76, 54, 3.21)
}
