% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{galactic}
\alias{galactic}
\title{Galactic frame}
\usage{
galactic()
}
\value{
A <sky_frame> object.
}
\description{
Galactic coordinate system, a celestial frame aligned with the Milky Way
plane, with longitude `l` and latitude `b`.
}
\references{
\url{https://en.wikipedia.org/wiki/Galactic_coordinate_system}
}
