% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_sky_coord_notation}
\alias{get_sky_coord_notation}
\title{Get sky_coord print notation}
\usage{
get_sky_coord_notation()
}
\value{
Current notation value. Defaults to "hmsdms".
}
\description{
Get sky_coord print notation
}
