% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexagesimal.R
\name{hms_to_deg}
\alias{hms_to_deg}
\title{Convert HMS components to degrees}
\usage{
hms_to_deg(h = 0, m = 0, s = 0)
}
\arguments{
\item{h}{Hours. Allowed range is `0 <= h < 24`, with one special case:
`h = 24` is allowed only when `m = 0` and `s = 0`.}

\item{m}{Minutes. Allowed range is `0 <= m < 60`.}

\item{s}{Seconds. Allowed range is `0 <= s < 60`.}
}
\value{
Numeric vector in degrees.
}
\description{
Convert hour-minute-second components to decimal degrees.
}
\examples{
hms_to_deg(12, 34, 56)

hms_to_deg(24.0)
}
