% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iau-name.R
\name{iau_name}
\alias{iau_name}
\title{Build IAU-style source names from sky coordinates}
\usage{
iau_name(x, prefix = NULL, epoch = "J", ra_digits = 1, dec_digits = 0)
}
\arguments{
\item{x}{A <sky_coord> vector.}

\item{prefix}{Optional string added before the epoch code.}

\item{epoch}{Epoch code, usually "J".}

\item{ra_digits}{Number of digits after the decimal point in RA seconds.}

\item{dec_digits}{Number of digits after the decimal point in Dec seconds.}
}
\value{
Character vector.
}
\description{
Build IAU-style source names from sky coordinates
}
\examples{
x <- parse_coord("J230631.0+155633")

iau_name(x)
iau_name(x, prefix = "SRGA J")
iau_name(x, ra_digits = 2, dec_digits = 1)

# Non-ICRS inputs are transformed internally
iau_name(transform_to(x, galactic()))
}
