% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{icrs}
\alias{icrs}
\title{ICRS frame}
\usage{
icrs()
}
\value{
A <sky_frame> object.
}
\description{
International Celestial Reference System (ICRS), the standard inertial
celestial frame with right ascension and declination axes.
}
\references{
\url{https://en.wikipedia.org/wiki/International_Celestial_Reference_System}
}
