% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sky-coord.R
\name{is.sky_coord}
\alias{is.sky_coord}
\title{Check whether an object is a sky_coord vector}
\usage{
is.sky_coord(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
`TRUE` if `x` inherits from `<sky_coord>`, otherwise `FALSE`.
}
\description{
Check whether an object is a sky_coord vector
}
