% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian-date.R
\name{jd_fromdate}
\alias{jd_fromdate}
\title{Convert date-time to Julian Date (JD)}
\usage{
jd_fromdate(x, scale = "UTC")
}
\arguments{
\item{x}{Date-time input as `POSIXct`, `POSIXlt`, or `Date`.

  `Date` values are interpreted as `00:00:00` in UTC.}

\item{scale}{Time scale used by ERFA. Only `"UTC"` is currently supported.}
}
\value{
Numeric vector with Julian Date values (in days).
}
\description{
Convert date-time to Julian Date (JD)
}
\details{
The fractional part of JD encodes time of day.

JD starts at noon, so calendar midnight corresponds to `.5` in JD.
}
\examples{
t <- as.POSIXct("2026-03-08 12:34:56", tz = "UTC")
jd_fromdate(t)

jd_fromdate(as.Date("2026-03-08"))
}
\seealso{
\code{\link{jd2greg}}, \code{\link{mjd2greg}}
}
