% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian-date.R
\name{jd_to_datetime}
\alias{jd_to_datetime}
\title{Convert Julian Date (JD) to POSIXct}
\usage{
jd_to_datetime(jd, tz = "UTC")
}
\arguments{
\item{jd}{Numeric vector of Julian Date values.}

\item{tz}{Time zone for the returned `POSIXct` vector.}
}
\value{
`POSIXct` vector.
}
\description{
Convert Julian Date (JD) to POSIXct
}
\details{
`tz` affects the returned clock representation, not the underlying moment.

This function currently uses ERFA with fixed `scale = "UTC"` internally.
}
\examples{
jd <- 2461107.5
jd_to_datetime(jd, tz = "UTC")
jd_to_datetime(jd, tz = "Europe/Moscow")

x <- as.POSIXct("2026-03-08 00:00:00", tz = "UTC")
jd_to_datetime(datetime_to_jd(x), tz = "UTC")
}
\seealso{
\code{\link{datetime_to_jd}}, \code{\link{mjd_to_datetime}}
}
