% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{l}
\alias{l}
\title{Get Galactic longitude (l)}
\usage{
l(x)
}
\arguments{
\item{x}{A <sky_coord> vector in Galactic frame.}
}
\value{
Numeric vector with Galactic longitude values in degrees.
}
\description{
Extract Galactic longitude values (in degrees) from Galactic coordinates.
}
\examples{
x <- gal_coord(c(120, 121), c(-30, -31))
l(x)
}
