% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leap-seconds.R
\name{leap_second_dates}
\alias{leap_second_dates}
\title{Leap-second dates (UTC)}
\usage{
leap_second_dates()
}
\value{
Date vector with leap-second change dates (effective at 00:00 UTC).
}
\description{
Return UTC dates where the TAI-UTC offset changes by one second.
}
\details{
Values are provided by ERFA.


The returned date is when the new UTC offset is already in effect
(00:00 UTC at the start of that date).

The first returned date is `1972-07-01` (the first leap second was inserted
at the end of `1972-06-30`).
}
\examples{
leap_second_dates()
}
