% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sky-coord.R
\name{new_sky_coord}
\alias{new_sky_coord}
\title{Low-level sky coordinate constructor}
\usage{
new_sky_coord(lon = double(), lat = double(), frame = icrs())
}
\arguments{
\item{lon, lat}{Longitude and latitude in degrees.}

\item{frame}{A sky frame object.}
}
\value{
A <sky_coord> vector.
}
\description{
Low-level sky coordinate constructor
}
