% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-coord.R
\name{parse_coord}
\alias{parse_coord}
\title{Parse coordinate strings into ICRS sky_coord}
\usage{
parse_coord(x = character())
}
\arguments{
\item{x}{Character vector with ICRS coordinates.}
}
\value{
A <sky_coord> vector in ICRS frame.
}
\description{
Parse coordinate strings into ICRS sky_coord
}
