% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-coord.R
\name{radec}
\alias{radec}
\title{Deprecated alias for parse_coord}
\usage{
radec(x = character())
}
\arguments{
\item{x}{Character vector with ICRS coordinates.}
}
\value{
A <sky_coord> vector in ICRS frame.
}
\description{
Use \code{\link{parse_coord}} instead.
}
\seealso{
\code{\link{parse_coord}}
}
