% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation.R
\name{separation}
\alias{separation}
\title{Angular separation between sky coordinates}
\usage{
separation(x, y)
}
\arguments{
\item{x, y}{<sky_coord> vectors.}
}
\value{
Numeric vector in arcseconds.
}
\description{
Angular separation between sky coordinates
}
\examples{
x <- ra_dec(10, 20)
y <- ra_dec(11, 21)
separation(x, y)

g <- transform_to(x, galactic())
separation(g, y)
}
