% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_print_hms}
\alias{set_print_hms}
\title{Use HMS/DMS print style for sky_coord}
\usage{
set_print_hms()
}
\value{
Invisibly returns the previous notation.
}
\description{
Use HMS/DMS print style for sky_coord
}
