% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_print_pair}
\alias{set_print_pair}
\title{Use pair print style for sky_coord}
\usage{
set_print_pair()
}
\value{
Invisibly returns the previous notation.
}
\description{
Use pair print style for sky_coord
}
