% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_print_plain}
\alias{set_print_plain}
\title{Use plain print style for sky_coord}
\usage{
set_print_plain()
}
\value{
Invisibly returns the previous notation.
}
\description{
Use plain print style for sky_coord
}
