% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_sky_coord_notation}
\alias{set_sky_coord_notation}
\title{Set sky_coord print notation}
\usage{
set_sky_coord_notation(notation = c("hmsdms", "plain", "pair"))
}
\arguments{
\item{notation}{One of "hmsdms", "plain", "pair".}
}
\value{
Invisibly returns the previous notation.
}
\description{
Set sky_coord print notation
}
