% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sky-coord.R
\name{sky_coord-vctrs-methods}
\alias{sky_coord-vctrs-methods}
\alias{vec_ptype_abbr.sky_coord}
\alias{vec_ptype_full.sky_coord}
\title{vctrs ptype methods for sky_coord}
\usage{
vec_ptype_abbr.sky_coord(x, ...)

vec_ptype_full.sky_coord(x, ...)
}
\arguments{
\item{x}{A <sky_coord> vector.}
}
\description{
Internal methods used by vctrs/pillar to display type abbreviations.
}
\keyword{internal}
