% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sky-coord.R
\name{sky_coord}
\alias{sky_coord}
\title{User-facing sky coordinate constructor}
\usage{
sky_coord(lon = double(), lat = double(), frame = icrs())
}
\arguments{
\item{lon, lat}{Longitude and latitude.}

\item{frame}{A sky frame object.}
}
\value{
A <sky_coord> vector.
}
\description{
User-facing sky coordinate constructor
}
