% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{sky_frame}
\alias{sky_frame}
\title{Generic sky frame constructor}
\usage{
sky_frame(name, x_name, y_name)
}
\arguments{
\item{name}{Frame name.}

\item{x_name}{Name of the longitude-like axis.}

\item{y_name}{Name of the latitude-like axis.}
}
\value{
A <sky_frame> object.
}
\description{
Generic sky frame constructor
}
