% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian.R
\name{to_cartesian}
\alias{to_cartesian}
\title{Convert sky coordinates to Cartesian unit vectors}
\usage{
to_cartesian(x)
}
\arguments{
\item{x}{A `<sky_coord>` vector.}
}
\value{
A base `data.frame` with numeric columns `x`, `y`, `z`.
}
\description{
Convert spherical sky coordinates (degrees) to unit Cartesian vectors.
}
