% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_galactic_to_icrs}
\alias{transform_galactic_to_icrs}
\title{Transform Galactic coordinates to ICRS frame}
\usage{
transform_galactic_to_icrs(x)
}
\arguments{
\item{x}{A <sky_coord> vector in Galactic frame.}
}
\value{
A <sky_coord> vector in ICRS frame.
}
\description{
Transform Galactic coordinates to ICRS frame
}
