% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_icrs_to_galactic}
\alias{transform_icrs_to_galactic}
\title{Transform ICRS coordinates to Galactic frame}
\usage{
transform_icrs_to_galactic(x)
}
\arguments{
\item{x}{A <sky_coord> vector in ICRS frame.}
}
\value{
A <sky_coord> vector in Galactic frame.
}
\description{
Transform ICRS coordinates to Galactic frame
}
