% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_to}
\alias{transform_to}
\title{Transform sky coordinates to another frame}
\usage{
transform_to(x, frame)
}
\arguments{
\item{x}{A <sky_coord> vector.}

\item{frame}{Target frame.}
}
\value{
A transformed <sky_coord> vector.
}
\description{
Transform sky coordinates to another frame
}
