% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astrocoords-package.R
\docType{package}
\name{astrocoords-package}
\alias{astrocoords}
\alias{astrocoords-package}
\title{astrocoords: Astronomical Coordinates for R}
\description{
Vectorized astronomical coordinates for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://uskovgs.github.io/astrocoords/}
}

}
\author{
\strong{Maintainer}: Grigory Uskov \email{uskov.russia@gmail.com} (\href{https://orcid.org/0000-0002-0274-1350}{ORCID})

}
\keyword{internal}
