% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian-date.R
\name{datetime_to_jd}
\alias{datetime_to_jd}
\title{Convert date-time to Julian Date (JD)}
\usage{
datetime_to_jd(x)
}
\arguments{
\item{x}{Date-time input as `POSIXct`, `POSIXlt`, or `Date`.

  `Date` values are interpreted as `00:00:00` in UTC.}
}
\value{
Numeric vector with Julian Date values (in days).
}
\description{
Convert date-time to Julian Date (JD)
}
\details{
The fractional part of JD encodes time of day.

JD starts at noon, so calendar midnight corresponds to `.5` in JD.

This function currently uses ERFA with fixed `scale = "UTC"` internally.
}
\examples{
old_accuracy <- getOption("digits")
options(digits = 11)

t <- as.POSIXct("2026-03-08 12:34:56", tz = "UTC")
datetime_to_jd(t)

datetime_to_jd(as.Date("2026-03-08"))

options(digits = old_accuracy)
}
\seealso{
\code{\link{jd_to_datetime}}, \code{\link{mjd_to_datetime}}
}
