% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{gal_coord}
\alias{gal_coord}
\title{Sugar constructor for Galactic coordinates}
\usage{
gal_coord(l, b)
}
\arguments{
\item{l, b}{Coordinates in degrees.}
}
\value{
A <sky_coord> vector in Galactic frame.
}
\description{
Sugar constructor for Galactic coordinates
}
