% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iau-name.R
\name{iau_name}
\alias{iau_name}
\title{Build IAU-style source names from sky coordinates}
\usage{
iau_name(x, prefix = NULL, epoch = "J", ra_digits = 1, dec_digits = 0)
}
\arguments{
\item{x}{A <sky_coord> vector.}

\item{prefix}{Optional string added before the epoch code.}

\item{epoch}{Epoch code, usually "J".}

\item{ra_digits}{Number of digits after the decimal point in RA seconds.}

\item{dec_digits}{Number of digits after the decimal point in Dec seconds.}
}
\value{
Character vector.
}
\description{
Build IAU-style source names from sky coordinates
}
