% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian-date.R
\name{jd2greg}
\alias{jd2greg}
\title{POSIXct from Julian Date}
\usage{
jd2greg(jd, scale = "UTC", tz = "UTC")
}
\arguments{
\item{jd}{Numeric vector of Julian Date.}

\item{scale}{ERFA timescale. Only "UTC" is currently supported.}

\item{tz}{Time zone for returned POSIXct.}
}
\value{
POSIXct vector.
}
\description{
POSIXct from Julian Date
}
