% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian-date.R
\name{jd_fromdate}
\alias{jd_fromdate}
\title{Julian Date from date-time}
\usage{
jd_fromdate(x, scale = "UTC")
}
\arguments{
\item{x}{POSIXct/POSIXlt or Date vector.}

\item{scale}{ERFA timescale. Only "UTC" is currently supported.}
}
\value{
Numeric vector with Julian Date.
}
\description{
Julian Date from date-time
}
