% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian-date.R
\name{mjd2greg}
\alias{mjd2greg}
\title{POSIXct from Modified Julian Date}
\usage{
mjd2greg(mjd, scale = "UTC", tz = "UTC")
}
\arguments{
\item{mjd}{Numeric vector of Modified Julian Date.}

\item{scale}{ERFA timescale. Only "UTC" is currently supported.}

\item{tz}{Time zone for returned POSIXct.}
}
\value{
POSIXct vector.
}
\description{
POSIXct from Modified Julian Date
}
