% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian-date.R
\name{mjd_to_datetime}
\alias{mjd_to_datetime}
\title{Convert Modified Julian Date (MJD) to POSIXct}
\usage{
mjd_to_datetime(mjd, tz = "UTC")
}
\arguments{
\item{mjd}{Numeric vector of Modified Julian Date values.}

\item{tz}{Time zone for the returned `POSIXct` vector.}
}
\value{
`POSIXct` vector.
}
\description{
Convert Modified Julian Date (MJD) to POSIXct
}
\details{
MJD is related to JD by:
\deqn{MJD = JD - 2400000.5}

This function currently uses ERFA with fixed `scale = "UTC"` internally.
}
\examples{
mjd_to_datetime(61107, tz = "UTC")

m <- 61107.25
mjd_to_datetime(m, tz = "UTC")
jd_to_datetime(m + 2400000.5, tz = "UTC")
}
\seealso{
\code{\link{jd_to_datetime}}, \code{\link{datetime_to_jd}}
}
