% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{ra_dec}
\alias{ra_dec}
\title{Create ICRS sky coordinates}
\usage{
ra_dec(ra, dec)
}
\arguments{
\item{ra, dec}{Coordinates in degrees.}
}
\value{
A <sky_coord> vector in ICRS frame.
}
\description{
Convenience wrapper for `sky_coord(..., frame = icrs())`.
}
