% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sky-coord.R
\name{sky_coord}
\alias{sky_coord}
\title{Create sky coordinates}
\usage{
sky_coord(lon = double(), lat = double(), frame = icrs())
}
\arguments{
\item{lon}{Longitude-like coordinate in degrees.

  For `frame = icrs()`, this is right ascension (`ra`) and must be between
  `0` and `360`.

  For `frame = galactic()`, this is Galactic longitude (`l`) and must be
  between `0` and `360`.}

\item{lat}{Latitude-like coordinate in degrees.

  For `frame = icrs()`, this is declination (`dec`) and must be between
  `-90` and `90`.

  For `frame = galactic()`, this is Galactic latitude (`b`) and must be
  between `-90` and `90`.}

\item{frame}{A sky frame object. Supported frames are
\code{\link[=icrs]{icrs()}} and \code{\link[=galactic]{galactic()}}.
The default is \code{\link[=icrs]{icrs()}}.}
}
\value{
A <sky_coord> vector.
}
\description{
Create a vector of sky coordinates in degrees.
}
\examples{
x <- sky_coord(10, 20)
x

g <- sky_coord(120, -30, frame = galactic())
g

ra_dec(10, 20)
gal_coord(120, -30)
}
